rename infer and check
